Updates scripts for external dependencies

+++ install/external/buildNamedCapture.sh
@@ -1,22 +1,22 @@
 #!/bin/bash -x
-# Run in Git for Windows Bash shell
-# You can just clone https://github.com:/tdhock/namedCapture into this
-# external folder in order to have something to build!
+# Run in Git for Windows Bash shell or equivalent
+# You should do this under Windows or you'll be stuck with a "source-only" build
 
-# Establish a temporary library in preparation for binary build
-TMP_LIBRARY="../tmp"
-echo $TMP_LIBRARY
-if [ ! -d $TMP_LIBRARY ]
-then
-	mkdir "$TMP_LIBRARY"
-fi
-echo "Installing into $TMP_LIBRARY"
+BUILT_DIR=../built
+EXT_DIR=$(pwd)
+TMP_LIBRARY="../tmp" # Same one used to hold products of building VisionEval packages
+
+# Lightly check that we have something to build and places to put it
+[ ! -d namedCapture ]   && git clone https://github.com:/tdhock/namedCapture.git "${EXT_DIR}/namedCapture"
+[ ! -d $BUILT_DIR/src ] && mkdir $BUILT_DIR/src
+[ ! -d $BUILT_DIR/bin ] && mkdir $BUILT_DIR/bin
+[ ! -d $TMP_LIBRARY ]   && mkdir $TMP_LIBRARY
 
 # Build source and binary packages for namedCapture
-cd ../built
-R CMD build ../external/namedCapture
-mv namedCapture*.tar.gz src
-R CMD INSTALL --build --library=$TMP_LIBRARY ../external/namedCapture
-mv namedCapture*.zip bin
+cd $BUILT_DIR
 
+R CMD build "${EXT_DIR}/namedCapture"
+[ -f namedCapture*.tar.gz ] && mv namedCapture*.tar.gz src
 
+R CMD INSTALL --build --library=$TMP_LIBRARY "${EXT_DIR}/namedCapture"
+[ -f namedCapture*.zip ] && mv namedCapture*.zip bin
